A little more stuff for a ROS dev container.
